Revert "Merge branch 'Benj1-mvc' into 158-mvc-refactoring"
This reverts commit 2f06dd65b126253aacd1b19100d7e333f64930ca, reversing
changes made to 98c750d555022f367dfbf118bc124c3ef74a86b0.
