+++ Frontend/vta_app/lib/src/ui/widgets/board/artifact.dart
