+++ Frontend/vta_app/lib/src/ui/widgets/categories/categories_widget.dart
@@ -8,7 +8,7 @@ import 'package:vta_app/src/modelsDTOs/artefact.dart';
 import 'package:vta_app/src/modelsDTOs/category.dart';
 import 'package:vta_app/src/notifiers/vta_notifiers.dart';
 import 'package:vta_app/src/ui/screens/take_picture_screen.dart';
-import 'package:vta_app/src/ui/widgets/board/artefact.dart';
+import 'package:vta_app/src/ui/widgets/board/artifact.dart';
 import 'package:vta_app/src/ui/widgets/board/talking_mat.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:vta_app/src/ui/widgets/categories/addPicture.dart';
@@ -17,20 +17,16 @@ import 'package:vta_app/src/ui/widgets/utilities/custom_delay_drag_listener.dart
 import 'package:vta_app/src/utilities/services/camera_service.dart';
 import 'package:http/http.dart' as http;
 
-import '../../../controllers/board_controller.dart';
-
 class CategoriesWidget extends StatefulWidget {
   final List<Category> categories;
   final double widgetHeight;
   final GlobalKey<TalkingMatState> talkingMatKey;
-  final BoardController boardController;
 
   const CategoriesWidget({
     super.key,
     required this.categories,
     required this.widgetHeight,
     required this.talkingMatKey,
-    required this.boardController
   });
 
   @override
@@ -38,30 +34,81 @@ class CategoriesWidget extends StatefulWidget {
 }
 
 class _CategoriesWidgetState extends State<CategoriesWidget> {
+  late List<Category> categories;
   late ArtifactState artifactState;
   late AuthState authState;
+  bool moveCategoriesMode = false;
 
   @override
   void initState() {
     super.initState();
     artifactState = Provider.of<ArtifactState>(context, listen: false);
     authState = Provider.of<AuthState>(context, listen: false);
+    categories = widget.categories;
   }
 
   @override
   Widget build(BuildContext context) {
-  final boardController = widget.boardController;
-
     return TapRegion(
         onTapOutside: (event) => {
-              if (boardController.moveCategoriesMode)
+              if (moveCategoriesMode)
                 {
                   setState(() {
-                    boardController.setMoveCategories(false);
+                    moveCategoriesMode = false;
                   })
                 }
             },
         child: _buildCategoryList());
+  }
+
+  Widget _buildCategoryList() {
+    return Theme(
+      data: ThemeData(
+        canvasColor: Colors.transparent,
+        shadowColor: Colors.transparent,
+      ),
+      child: ReorderableListView.builder(
+        scrollDirection: Axis.horizontal,
+        buildDefaultDragHandles: false,
+        itemCount: categories.length + 1, //+1 room for add button
+        itemBuilder: (context, index) {
+          if (index == categories.length) {
+            return _buildAddCategoryButton(key: ValueKey('add_button'));
+          }
+          if (moveCategoriesMode) {
+            return Material(
+              key: ValueKey(categories[index].categoryId),
+              elevation: 2,
+              child: _buildCategoryItem(context, index),
+            );
+          }
+          return _buildCategoryItem(context, index,
+              key: ValueKey(categories[index].categoryId));
+        },
+        onReorder: (int oldIndex, int newIndex) {
+          setState(() {
+            if (oldIndex < newIndex) {
+              newIndex -= 1;
+            }
+            newIndex = min(newIndex, categories.length - 1);
+
+            final Category movedCategory = categories.removeAt(oldIndex);
+            categories.insert(newIndex, movedCategory);
+
+            for (int i = min(oldIndex, newIndex);
+                i <= max(oldIndex, newIndex);
+                i++) {
+              categories[i].categoryIndex = i;
+              categories[i].userId = authState.userId;
+              artifactState.updateCategory(
+                categories[i],
+                token: authState.token!,
+              );
+            }
+          });
+        },
+      ),
+    );
   }
 
   Widget _buildAddCategoryButton({Key? key}) {
@@ -98,10 +145,8 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
   }
 
   Widget _buildCategoryItem(BuildContext context, int index, {Key? key}) {
-    final boardController = widget.boardController;
-
-    final item = boardController.categories[index];
-    if (boardController.moveCategoriesMode) {
+    final item = categories[index];
+    if (moveCategoriesMode) {
       return CustomDelayDragStartListener(
         delay: 200,
         key: key,
@@ -117,6 +162,15 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
         child: _buildCategoryButton(key, context, index, item),
       );
     }
+  }
+
+  TextButton _buildCategoryButton(
+      Key? key, BuildContext context, int index, Category item) {
+    return TextButton(
+      key: key,
+      onPressed: () => _showCategoryModal(context, categories[index]),
+      child: _buildCategoryContainer(item),
+    );
   }
 
   Widget _buildCategoryContainer(Category item) {
@@ -220,7 +274,7 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
                     const Text('Move', style: TextStyle(color: Colors.green)),
                 onTap: () {
                   setState(() {
-                    boardController.setMoveCategories(true);
+                    moveCategoriesMode = true;
                     Navigator.pop(context); // Close the modal
                   });
                 },
@@ -301,6 +355,7 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
       );
     });
   }
+
   Widget _buildImageGridItem(BuildContext context, int index, Category category,
       bool isInDeletionMode, VoidCallback onLongPress,
       {required VoidCallback onDelete}) {
@@ -309,7 +364,7 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
     var headers = <String, String>{
       'Authorization': 'Bearer ${authState.token}'
     };
-    final boardController = Provider.of<BoardController>(context);
+
     if (index >= category.artefacts!.length) {
       return SizedBox(); // Safety check
     }
@@ -318,6 +373,7 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
         .map((artefact) =>
             BoardArtefact.fromArtefact(artefact, headers: headers))
         .toList();
+
     return GestureDetector(
       onLongPress: onLongPress,
       child: Stack(
@@ -328,7 +384,8 @@ class _CategoriesWidgetState extends State<CategoriesWidget> {
               onPressed: isInDeletionMode
                   ? null
                   : () {
-                boardController.addArtefact(boardArtefacts[index]);
+                      widget.talkingMatKey.currentState
+                          ?.addArtifact(boardArtefacts[index]);
                       Navigator.pop(context);
                     },
               child: boardArtefacts[index].content,
